Update for bugfixes and file sizes
